Rename AmazonBraketTask to BraketTask
This makes all the provider class names consistent
