+++ qiskit_braket_provider/providers/braket_backend.py
@@ -25,7 +25,7 @@ from .adapter import (
     local_simulator_to_target,
     to_braket,
 )
-from .braket_job import AmazonBraketTask
+from .braket_task import BraketTask
 
 logger = logging.getLogger(__name__)
 
@@ -125,7 +125,7 @@ class BraketLocalBackend(BraketBackend):
 
     def run(
         self, run_input: Union[QuantumCircuit, list[QuantumCircuit]], **options
-    ) -> AmazonBraketTask:
+    ) -> BraketTask:
         convert_input = (
             [run_input] if isinstance(run_input, QuantumCircuit) else list(run_input)
         )
@@ -160,7 +160,7 @@ class BraketLocalBackend(BraketBackend):
 
         task_id = _TASK_ID_DIVIDER.join(task.id for task in tasks)
 
-        return AmazonBraketTask(
+        return BraketTask(
             task_id=task_id,
             tasks=tasks,
             backend=self,
@@ -220,7 +220,7 @@ class BraketAwsBackend(BraketBackend):
         )
         self._target = aws_device_to_target(device=device)
 
-    def retrieve_job(self, task_id: str) -> AmazonBraketTask:
+    def retrieve_job(self, task_id: str) -> BraketTask:
         """Return a single job submitted to AWS backend.
 
         Args:
@@ -231,7 +231,7 @@ class BraketAwsBackend(BraketBackend):
         """
         task_ids = task_id.split(_TASK_ID_DIVIDER)
 
-        return AmazonBraketTask(
+        return BraketTask(
             task_id=task_id,
             backend=self,
             tasks=[AwsQuantumTask(arn=task_id) for task_id in task_ids],
@@ -336,7 +336,7 @@ class BraketAwsBackend(BraketBackend):
         tasks: list[AwsQuantumTask] = batch_task.tasks
         task_id = _TASK_ID_DIVIDER.join(task.id for task in tasks)
 
-        return AmazonBraketTask(
+        return BraketTask(
             task_id=task_id, tasks=tasks, backend=self, shots=options.get("shots")
         )
